% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{ttss_summary}
\alias{ttss_summary}
\title{second structure summary}
\usage{
ttss_summary(str, file_paired, file_unpaired)
}
\arguments{
\item{str}{second structure string}

\item{file_paired}{output file paired}

\item{file_unpaired}{output file unpaired}
}
\description{
second structure summary
}
